services/voice-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
replicate==0.34.0
pydub==0.25.1
ffmpeg-python==0.2.0
boto3==1.34.120
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2